Fixed fp16 issue

+++ image_model.py
@@ -41,7 +41,8 @@ class ModifiedResnet(models.ResNet):
         fcnn_results = self.fcnn(x)
         max_feature_nums = max(map(lambda x: x["boxes"].size(0), fcnn_results))
         feat_dim = fcnn_results[0]["features"].size(-1)
-        features = torch.zeros((len(fcnn_results), max_feature_nums, feat_dim + 7)).to(location_embedding.device)
+        features = torch.zeros((len(fcnn_results), max_feature_nums, feat_dim + 7), dtype=location_embedding.dtype).to(
+            location_embedding.device)
         object_labels = torch.zeros((len(fcnn_results), max_feature_nums), dtype=torch.long).to(
             location_embedding.device)
         for i in range(len(fcnn_results)):
